+++ Python/hbayesdm/models/_rdt_happiness.py
@@ -1,7 +1,7 @@
 from typing import Sequence, Union, Any
 from collections import OrderedDict
 
-from numpy import Inf, exp
+from numpy import inf, exp
 import pandas as pd
 
 from hbayesdm.base import TaskModel
@@ -28,12 +28,12 @@ class RdtHappiness(TaskModel):
                 'RT_happy',
             ),
             parameters=OrderedDict([
-                ('w0', (-Inf, 1, Inf)),
-                ('w1', (-Inf, 1, Inf)),
-                ('w2', (-Inf, 1, Inf)),
-                ('w3', (-Inf, 1, Inf)),
+                ('w0', (-inf, 1, inf)),
+                ('w1', (-inf, 1, inf)),
+                ('w2', (-inf, 1, inf)),
+                ('w3', (-inf, 1, inf)),
                 ('gam', (0, 0.5, 1)),
-                ('sig', (0, 1, Inf)),
+                ('sig', (0, 1, inf)),
             ]),
             regressors=OrderedDict([