+++ R/R/preprocess_funcs.R
@@ -818,7 +818,7 @@ pst_preprocess_func <- function(raw_data, general_info) {
 }
 
 # Make a function
-pstRT_preprocess_func <- function(raw_data, general_info, RTbound, initQ) {
+pstRT_preprocess_func <- function(raw_data, general_info, RTbound = 0.1, initQ = 0.5) {
   # Use raw_data as a data.frame
   raw_data <- as.data.frame(raw_data)