+++ R/R/settings.R
@@ -1,6 +1,3 @@
 #' @noRd
-if (Sys.getenv('BUILD_ALL') == "true") {
-  FLAG_BUILD_ALL <- TRUE
-} else {
-  FLAG_BUILD_ALL <- FALSE
-}
+#' Reserved for package-level settings.
+NULL